refactor(linker): split lib.rs into focused submodules
aube-linker/src/lib.rs was 4662 lines. Carved out self-contained
helpers so lib.rs is now 2757 lines (-41%) and each file has a single
concern. No behaviour change — every call site keeps working via root
re-exports.

- error.rs: Error enum + diagnostic codes
- pool.rs: rayon thread-pool cache + default_linker_parallelism
- sweep.rs: tmp/top-level/hidden-hoist sweep + EntryState + classify
- patches.rs: Patches alias, patch sidecar I/O, diff apply + parse
- tests.rs / public_hoist_tests.rs: extracted top-level test modules
- importer + patch tests moved next to the code under test

Also deduped Linker::new — it now delegates to new_with_gvs (was an
identical 22-line body inlined twice).
