+++ crates/aube-linker/src/lib.rs
@@ -9,315 +9,32 @@ use aube_store::{PackageIndex, Store, StoredFile};
 use std::collections::BTreeMap;
 use std::path::{Path, PathBuf};
 
+mod error;
 mod hoisted;
+mod patches;
+mod pool;
+mod sweep;
 pub mod sys;
+
+#[cfg(test)]
+mod public_hoist_tests;
+#[cfg(test)]
+mod tests;
+
+pub use error::Error;
 pub use hoisted::HoistedPlacements;
-
-/// Sweep orphan `.tmp-<pid>-*` directories in the virtual store.
-///
-/// Linker materializes each package into `.tmp-<pid>-<subdir>/`
-/// then atomic-renames into `.aube/<subdir>/`. Crash or Ctrl-C
-/// between materialize and rename leaves the tmp dir behind.
-/// Nothing else cleans these up so they accumulate on every aborted
-/// install. Small footprint per entry but a few hundred aborted
-/// CI runs pile up gigabytes.
-///
-/// Called early in link_all so each fresh install reclaims space
-/// from prior crashes. Only matches the exact prefix we produce so
-/// user files named `.tmp-*` in the virtual store are safe.
-pub fn sweep_stale_tmp_dirs(virtual_store: &Path) {
-    let Ok(entries) = std::fs::read_dir(virtual_store) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let name = entry.file_name();
-        let name = name.to_string_lossy();
-        // Match our exact prefix. Format: `.tmp-<pid>-<subdir>`
-        // where pid is numeric.
-        if !name.starts_with(".tmp-") {
-            continue;
-        }
-        let rest = &name[".tmp-".len()..];
-        let Some((pid_str, _rest)) = rest.split_once('-') else {
-            continue;
-        };
-        if pid_str.chars().any(|c| !c.is_ascii_digit()) {
-            continue;
-        }
-        // Do not touch the dir of our own still-running process.
-        // Materialize path creates and removes its tmp dir in the
-        // same call and crashes mid-way are the target here, the
-        // active pid will not leave ones around that matter.
-        if pid_str == std::process::id().to_string() {
-            continue;
-        }
-        let _ = remove_dir_all_with_retry(&entry.path());
-    }
-}
-
-/// Remove a directory with retry on Windows sharing violations.
-///
-/// Windows does not let you delete a file while another process holds
-/// a handle open. Dev server, vitest watcher, tsc --watch all hold
-/// .js / .node files inside node_modules. aube reinstall hits ERROR
-/// 32 (SHARING_VIOLATION) or ERROR 5 (ACCESS_DENIED, AV scanner
-/// mid-scan) and leaves a half-deleted virtual store. pnpm, npm,
-/// rimraf all retry with backoff. Do the same. Unix passthrough.
-///
-/// Retries 10 times with exponential backoff starting at 50ms. Total
-/// worst case around 10 seconds which is tolerable for an install
-/// already paying for filesystem work.
-pub fn remove_dir_all_with_retry(path: &Path) -> std::io::Result<()> {
-    #[cfg(not(windows))]
-    {
-        std::fs::remove_dir_all(path)
-    }
-    #[cfg(windows)]
-    {
-        use std::io::ErrorKind;
-        let mut delay_ms = 50u64;
-        for attempt in 0..10 {
-            match std::fs::remove_dir_all(path) {
-                Ok(()) => return Ok(()),
-                Err(e) if e.kind() == ErrorKind::NotFound => return Ok(()),
-                Err(e) => {
-                    // Sharing violation and PermissionDenied both
-                    // map to retriable Windows errors. Bail on
-                    // attempt 10.
-                    let retriable =
-                        matches!(e.kind(), ErrorKind::PermissionDenied | ErrorKind::Other)
-                            || e.raw_os_error() == Some(32);
-                    if !retriable || attempt == 9 {
-                        return Err(e);
-                    }
-                    std::thread::sleep(std::time::Duration::from_millis(delay_ms));
-                    delay_ms = (delay_ms * 2).min(2000);
-                }
-            }
-        }
-        // Unreachable, loop always returns by attempt 10.
-        Ok(())
-    }
-}
-
-/// Real workspace importer, not a peer-context bookkeeping entry.
-///
-/// pnpm v9 lockfiles record the peer-resolution view of each
-/// workspace package reached through every nested `node_modules/`
-/// traversal. Those virtual importer paths (e.g.
-/// `packages/a/node_modules/@scope/b/node_modules/@scope/c`) describe
-/// *how* a package looks from a particular context — they are reached
-/// via the workspace-to-workspace symlink chain and have no
-/// independent `node_modules/` to populate. When the link pipeline
-/// treats them as physical importers it queues parallel symlink tasks
-/// whose `link_path`s canonicalize to the same inode as a physical
-/// importer's task, producing EEXIST races on large monorepos.
-pub fn is_physical_importer(importer_path: &str) -> bool {
-    importer_path == "." || !importer_path.contains("/node_modules/")
-}
-
-/// Wipe `path` when it looks like a linker-managed `.aube/node_modules`
-/// tree. If a previously-tampered install (or attacker) replaced the
-/// tree with a symlink / junction pointing elsewhere on disk, refuse
-/// to recurse into it — modern Rust `remove_dir_all` already declines
-/// to follow symlinks, mirroring the invariant at the call site keeps
-/// the intent explicit and catches any future regression in the
-/// callee.
-fn remove_hidden_hoist_tree(path: &Path) {
-    match std::fs::symlink_metadata(path) {
-        Ok(md) if md.file_type().is_symlink() => {
-            let _ = std::fs::remove_file(path);
-        }
-        Ok(_) => {
-            let _ = std::fs::remove_dir_all(path);
-        }
-        Err(_) => {}
-    }
-}
-
-/// Best-effort unlink of `path` regardless of whether it's a file,
-/// symlink, junction, or directory. Errors are intentionally ignored
-/// because this is a "clear the slot" operation — the caller is about
-/// to place something else here and any residual entry that survives
-/// will surface as a downstream error.
-pub(crate) fn try_remove_entry(path: &Path) {
-    let _ = std::fs::remove_dir_all(path);
-    let _ = std::fs::remove_file(path);
-}
-
-/// `xx::file::mkdirp` wrapped with the linker's `Error::Xx` conversion.
-/// Every materialize pass calls this before creating a symlink /
-/// junction, so the lossy `.to_string()` wrap lives in exactly one
-/// place.
-pub fn mkdirp(dir: &Path) -> Result<(), Error> {
-    xx::file::mkdirp(dir).map_err(|e| Error::Xx(e.to_string()))
-}
-
-/// Classification of a `.aube/<dep_path>` symlink relative to the
-/// current hashed global entry the linker wants to point at.
-#[derive(Copy, Clone)]
-pub(crate) enum EntryState {
-    /// The symlink already points at `expected` and the target exists —
-    /// nothing to do. Caller can bump a `packages_cached` counter and
-    /// move on.
-    Fresh,
-    /// No entry at `link_path` yet. Caller needs to materialize and
-    /// create the symlink, but there's nothing to unlink first.
-    Missing,
-    /// An entry exists but is stale (different target, dangling link,
-    /// or an `Err` read that isn't NotFound). Caller must unlink
-    /// before resymlinking.
-    Stale,
-}
-
-/// Sweep stale entries out of a `node_modules/` directory while
-/// preserving everything in `preserve` (bare names like `lodash` and
-/// scope prefixes like `@babel`), dotfiles, and — if set — the
-/// virtual-store leaf (`aube_dir_leaf`) sitting right under `nm`
-/// with a non-dotfile name (the `virtualStoreDir=node_modules/vstore`
-/// case). For `@scope` entries we recurse one level and drop any
-/// `@scope/<pkg>` whose full `@scope/pkg` name is not in `preserve`;
-/// an empty scope directory left behind by the sweep is removed so
-/// the next install doesn't trip over a phantom scope tombstone.
-pub(crate) fn sweep_stale_top_level_entries(
-    nm: &Path,
-    preserve: &std::collections::HashSet<&str>,
-    aube_dir_leaf: Option<&std::ffi::OsStr>,
-) {
-    let scope_prefixes: std::collections::HashSet<&str> = preserve
-        .iter()
-        .filter_map(|n| n.split_once('/').map(|(scope, _)| scope))
-        .collect();
-    let Ok(entries) = std::fs::read_dir(nm) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let name = entry.file_name();
-        let name_str = name.to_string_lossy();
-        if name_str.starts_with('.') {
-            continue;
-        }
-        if aube_dir_leaf == Some(name.as_os_str()) {
-            continue;
-        }
-        if preserve.contains(name_str.as_ref()) {
-            continue;
-        }
-        if scope_prefixes.contains(name_str.as_ref()) {
-            let scope_dir = entry.path();
-            if let Ok(inner) = std::fs::read_dir(&scope_dir) {
-                for inner_entry in inner.flatten() {
-                    let inner_name = inner_entry.file_name();
-                    let full = format!("{}/{}", name_str, inner_name.to_string_lossy());
-                    if !preserve.contains(full.as_str()) {
-                        try_remove_entry(&inner_entry.path());
-                    }
-                }
-            }
-            // If the scope dir is now empty (every member was stale),
-            // drop the tombstone directory too.
-            if std::fs::read_dir(&scope_dir)
-                .map(|mut d| d.next().is_none())
-                .unwrap_or(false)
-            {
-                let _ = std::fs::remove_dir(&scope_dir);
-            }
-            continue;
-        }
-        try_remove_entry(&entry.path());
-    }
-}
-
-/// Sweep broken entries from a shared hidden-hoist directory without
-/// deleting live links owned by other projects. The GVS hidden hoist is
-/// global, so "not in this project's graph" is not stale enough: another
-/// project may still need that link. Only entries whose target no longer
-/// exists (or non-link junk) are reclaimed here; current-project names are
-/// still target-reconciled by `reconcile_top_level_link` below.
-pub(crate) fn sweep_dead_hidden_hoist_entries(hidden: &Path) {
-    let Ok(entries) = std::fs::read_dir(hidden) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let name = entry.file_name();
-        let name_str = name.to_string_lossy();
-        if name_str.starts_with('.') {
-            continue;
-        }
-        let path = entry.path();
-        if name_str.starts_with('@') {
-            match std::fs::symlink_metadata(&path) {
-                Ok(md) if md.is_dir() && !md.file_type().is_symlink() => {
-                    sweep_dead_hidden_hoist_scope(&path);
-                    if std::fs::read_dir(&path)
-                        .map(|mut d| d.next().is_none())
-                        .unwrap_or(false)
-                    {
-                        let _ = std::fs::remove_dir(&path);
-                    }
-                }
-                Ok(_) => sweep_dead_hidden_hoist_entry(&path),
-                Err(_) => {}
-            }
-            continue;
-        }
-        sweep_dead_hidden_hoist_entry(&path);
-    }
-}
-
-fn sweep_dead_hidden_hoist_scope(scope_dir: &Path) {
-    let Ok(entries) = std::fs::read_dir(scope_dir) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let name = entry.file_name();
-        if name.to_string_lossy().starts_with('.') {
-            continue;
-        }
-        sweep_dead_hidden_hoist_entry(&entry.path());
-    }
-}
-
-fn sweep_dead_hidden_hoist_entry(path: &Path) {
-    match std::fs::symlink_metadata(path) {
-        Ok(md) if md.file_type().is_symlink() && path.exists() => {}
-        Ok(md) if md.file_type().is_symlink() => {
-            try_remove_entry(path);
-        }
-        Ok(md) if md.is_dir() => {
-            try_remove_entry(path);
-        }
-        Ok(_) => {
-            try_remove_entry(path);
-        }
-        Err(_) => {}
-    }
-}
-
-/// Classify `link_path` against `expected` without the double-check
-/// (`read_link` then `exists`) that ate ~1.4k ENOENT syscalls per
-/// install on the medium fixture. Fresh means "points at expected
-/// AND the target still exists"; everything else is Missing or
-/// Stale. The fast path returns without touching disk a second time.
-#[inline]
-pub(crate) fn classify_entry_state(link_path: &Path, expected: &Path) -> EntryState {
-    match std::fs::read_link(link_path) {
-        Ok(existing) if existing == expected => {
-            if link_path.exists() {
-                EntryState::Fresh
-            } else {
-                EntryState::Stale
-            }
-        }
-        Ok(_) => EntryState::Stale,
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => EntryState::Missing,
-        // Some other error (permission, etc.): treat as Stale and
-        // let the removal/recreate path try its best-effort cleanup
-        // + surface the real error on symlink creation if unlucky.
-        Err(_) => EntryState::Stale,
-    }
-}
-
+pub use patches::Patches;
+pub(crate) use patches::{
+    apply_multi_file_patch, current_patch_hashes, read_applied_patches,
+    wipe_changed_patched_entries, write_applied_patches,
+};
+pub use pool::default_linker_parallelism;
+pub(crate) use pool::with_link_pool;
+pub(crate) use sweep::{
+    EntryState, classify_entry_state, remove_hidden_hoist_tree, sweep_dead_hidden_hoist_entries,
+    sweep_stale_top_level_entries, try_remove_entry,
+};
+pub use sweep::{is_physical_importer, mkdirp, remove_dir_all_with_retry, sweep_stale_tmp_dirs};
 pub use sys::{
     BinShimOptions, create_bin_shim, create_dir_link, normalize_path, parse_posix_shim_target,
     remove_bin_shim, validate_bin_name, validate_bin_target,
@@ -492,53 +209,6 @@ pub struct Linker {
     virtual_store_only: bool,
 }
 
-/// Patches to apply at materialize time, keyed by `name@version`. Each
-/// value is the raw multi-file unified diff text written by `aube
-/// patch-commit` (or any compatible tool).
-pub type Patches = std::collections::BTreeMap<String, String>;
-
-pub fn default_linker_parallelism() -> usize {
-    let default_limit = if cfg!(target_os = "macos") { 4 } else { 16 };
-
-    std::thread::available_parallelism()
-        .map(|n| n.get())
-        .unwrap_or(1)
-        .min(default_limit)
-}
-
-type LinkPoolCache = std::sync::Mutex<Vec<(usize, std::sync::Arc<rayon::ThreadPool>)>>;
-static LINK_POOL_CACHE: std::sync::OnceLock<LinkPoolCache> = std::sync::OnceLock::new();
-
-fn link_pool(threads: usize) -> Option<std::sync::Arc<rayon::ThreadPool>> {
-    let cache = LINK_POOL_CACHE.get_or_init(|| std::sync::Mutex::new(Vec::new()));
-    let mut guard = cache.lock().ok()?;
-    if let Some((_, pool)) = guard.iter().find(|(t, _)| *t == threads) {
-        return Some(pool.clone());
-    }
-    match rayon::ThreadPoolBuilder::new()
-        .num_threads(threads)
-        .thread_name(|i| format!("aube-linker-{i}"))
-        .build()
-    {
-        Ok(pool) => {
-            let pool = std::sync::Arc::new(pool);
-            guard.push((threads, pool.clone()));
-            Some(pool)
-        }
-        Err(err) => {
-            warn!("failed to build aube linker thread pool: {err}; falling back to caller thread");
-            None
-        }
-    }
-}
-
-fn with_link_pool<R: Send>(threads: usize, f: impl FnOnce() -> R + Send) -> R {
-    match link_pool(threads) {
-        Some(pool) => pool.install(f),
-        None => f(),
-    }
-}
-
 /// Strategy for linking files from the store to node_modules.
 #[derive(Debug, Clone, Copy)]
 pub enum LinkStrategy {
@@ -555,29 +225,7 @@ pub enum LinkStrategy {
 
 impl Linker {
     pub fn new(store: &Store, strategy: LinkStrategy) -> Self {
-        let use_global_virtual_store = !aube_util::env::is_ci();
-        Self {
-            virtual_store: store.virtual_store_dir(),
-            store: store.clone(),
-            use_global_virtual_store,
-            strategy,
-            patches: Patches::new(),
-            hashes: None,
-            virtual_store_dir_max_length: DEFAULT_VIRTUAL_STORE_DIR_MAX_LENGTH,
-            shamefully_hoist: false,
-            public_hoist_patterns: Vec::new(),
-            public_hoist_negations: Vec::new(),
-            hoist: true,
-            hoist_patterns: vec![glob::Pattern::new("*").expect("'*' is a valid glob pattern")],
-            hoist_negations: Vec::new(),
-            hoist_workspace_packages: true,
-            dedupe_direct_deps: false,
-            node_linker: NodeLinker::Isolated,
-            link_concurrency: None,
-            virtual_store_only: false,
-            modules_dir_name: "node_modules".to_string(),
-            aube_dir_override: None,
-        }
+        Self::new_with_gvs(store, strategy, !aube_util::env::is_ci())
     }
 
     /// Select the layout mode. Defaults to `NodeLinker::Isolated`
@@ -898,7 +546,6 @@ impl Linker {
         self.use_global_virtual_store
     }
 
-    #[cfg(test)]
     fn new_with_gvs(store: &Store, strategy: LinkStrategy, use_global_virtual_store: bool) -> Self {
         Self {
             virtual_store: store.virtual_store_dir(),
@@ -2931,36 +2578,6 @@ pub struct LinkStats {
     pub hoisted_placements: Option<HoistedPlacements>,
 }
 
-#[derive(Debug, thiserror::Error, miette::Diagnostic)]
-pub enum Error {
-    #[error("I/O error at {0}: {1}")]
-    Io(PathBuf, std::io::Error),
-    #[error("file error: {0}")]
-    Xx(String),
-    #[error("failed to link {0} -> {1}: {2}")]
-    #[diagnostic(code(ERR_AUBE_LINK_FAILED))]
-    Link(PathBuf, PathBuf, String),
-    #[error("failed to apply patch for {0}: {1}")]
-    #[diagnostic(code(ERR_AUBE_PATCH_FAILED))]
-    Patch(String, String),
-    #[error(
-        "internal: missing package index for {0} — caller skipped `load_index` but the package wasn't already materialized"
-    )]
-    #[diagnostic(code(ERR_AUBE_MISSING_PACKAGE_INDEX))]
-    MissingPackageIndex(String),
-    #[error("refusing to materialize unsafe index key: {0:?}")]
-    #[diagnostic(code(ERR_AUBE_UNSAFE_INDEX_KEY))]
-    UnsafeIndexKey(String),
-    #[error(
-        "cached package index references a missing CAS shard at {store_path} (file: {rel_path:?}). The store and its index cache are out of sync — rerun the install to re-fetch the tarball."
-    )]
-    #[diagnostic(code(ERR_AUBE_MISSING_STORE_FILE))]
-    MissingStoreFile {
-        store_path: PathBuf,
-        rel_path: String,
-    },
-}
-
 /// Defence in depth for the tarball path-traversal class. The
 /// primary guard lives in `aube_store::import_tarball`, which
 /// refuses malformed entries before they enter the `PackageIndex`.
@@ -3120,21 +2737,6 @@ fn reconcile_top_level_link(link_path: &Path, expected_target: &Path) -> Result<
     }
 }
 
-/// Compute per-`(name@version)` content hashes for the currently
-/// configured patch set. Returns a stable map so the caller can
-/// compare it against a sidecar from a previous install.
-fn current_patch_hashes(patches: &Patches) -> std::collections::BTreeMap<String, String> {
-    use sha2::{Digest, Sha256};
-    patches
-        .iter()
-        .map(|(k, v)| {
-            let mut h = Sha256::new();
-            h.update(v.as_bytes());
-            (k.clone(), hex::encode(h.finalize()))
-        })
-        .collect()
-}
-
 /// Build a `dep_path → absolute on-disk target` map for every
 /// `LocalSource::Link` in the graph. Returned `None` when the graph
 /// has no link entries (vast majority of installs), so the materialize
@@ -3152,1534 +2754,4 @@ pub fn build_nested_link_targets(
         })
         .collect();
     if map.is_empty() { None } else { Some(map) }
-}
-
-/// Read the previously-applied patch sidecar at
-/// `node_modules/.aube-applied-patches.json`. Missing or malformed
-/// files return an empty map — the caller treats them as "no patches
-/// were ever applied here," which conservatively triggers a re-link
-/// on the first run after the linker started writing the sidecar.
-fn read_applied_patches(nm_dir: &Path) -> std::collections::BTreeMap<String, String> {
-    let path = nm_dir.join(".aube-applied-patches.json");
-    let Ok(raw) = std::fs::read_to_string(&path) else {
-        return Default::default();
-    };
-    serde_json_parse_map(&raw).unwrap_or_default()
-}
-
-/// Tiny hand-rolled JSON object parser specialized for the sidecar:
-/// `{"name@ver": "hex", ...}`. Avoids dragging serde_json into
-/// `aube-linker` for one file. Returns `None` on any malformed input
-/// so the caller falls back to "no previous state."
-fn serde_json_parse_map(s: &str) -> Option<std::collections::BTreeMap<String, String>> {
-    // Old code hand-rolled JSON with split(',') and trim_matches('"').
-    // Breaks on any key or value containing a comma, escaped quote,
-    // or backslash. Keys are name@version strings today but patch
-    // content hashes are fine, and if pnpm ever extends the key
-    // schema this blows up. Real JSON parser handles escaping.
-    serde_json::from_str(s).ok()
-}
-
-/// Write the applied-patch sidecar.
-///
-/// Next install reads this to compute which `.aube/<dep_path>`
-/// entries need re-materializing because their patch set changed.
-/// Old code was `let _ = fs::write(...)`, dropped any IO error. If
-/// write silently failed (disk full, read-only mount, perms), the
-/// sidecar was missing on next install, and
-/// wipe_changed_patched_entries did not know which entries to
-/// re-link. Install reported success while node_modules had stale
-/// patched content on disk. Return Result, caller logs loudly.
-fn write_applied_patches(
-    nm_dir: &Path,
-    map: &std::collections::BTreeMap<String, String>,
-) -> std::io::Result<()> {
-    let path = nm_dir.join(".aube-applied-patches.json");
-    let out = serde_json::to_string(map)
-        .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-    aube_util::fs_atomic::atomic_write(&path, out.as_bytes())
-}
-
-/// Wipe `.aube/<dep_path>` for any package whose patch fingerprint
-/// changed between the previous and current install. Used by the
-/// per-project (no-global-store) link path, where the directory name
-/// doesn't otherwise change when a patch is added or removed.
-fn wipe_changed_patched_entries(
-    aube_dir: &Path,
-    graph: &LockfileGraph,
-    prev: &std::collections::BTreeMap<String, String>,
-    curr: &std::collections::BTreeMap<String, String>,
-    max_length: usize,
-) {
-    let mut affected: std::collections::HashSet<String> = std::collections::HashSet::new();
-    for k in prev.keys().chain(curr.keys()) {
-        if prev.get(k) != curr.get(k) {
-            affected.insert(k.clone());
-        }
-    }
-    if affected.is_empty() {
-        return;
-    }
-    for (dep_path, pkg) in &graph.packages {
-        let key = pkg.spec_key();
-        if affected.contains(&key) {
-            let entry = aube_dir.join(dep_path_to_filename(dep_path, max_length));
-            let _ = std::fs::remove_dir_all(entry);
-        }
-    }
-}
-
-/// Apply a git-style multi-file unified diff to a package directory.
-///
-/// The patch text is split on `diff --git ` boundaries; each section
-/// is parsed as a single-file unified diff and applied to the matching
-/// file under `pkg_dir`. We deliberately unlink the destination
-/// before writing, because the linker materializes files via reflink
-/// or hardlink — modifying the file in place would corrupt the global
-/// content-addressed store the linked file points to.
-fn is_safe_rel_component(rel: &str) -> bool {
-    if rel.is_empty() || rel.contains('\0') || rel.contains('\\') {
-        return false;
-    }
-    let p = Path::new(rel);
-    if p.is_absolute()
-        || p.has_root()
-        || rel.starts_with('/')
-        || rel.len() >= 2 && rel.as_bytes()[1] == b':'
-    {
-        return false;
-    }
-    p.components().all(|c| {
-        matches!(
-            c,
-            std::path::Component::Normal(_) | std::path::Component::CurDir
-        )
-    })
-}
-
-fn ensure_no_symlink_in_chain(pkg_dir: &Path, rel: &str) -> Result<(), String> {
-    let mut cursor = pkg_dir.to_path_buf();
-    for comp in Path::new(rel).components() {
-        cursor.push(comp);
-        match std::fs::symlink_metadata(&cursor) {
-            Ok(meta) => {
-                if meta.file_type().is_symlink() {
-                    return Err(format!("{}", cursor.display()));
-                }
-                // Junctions on Windows are `IO_REPARSE_TAG_MOUNT_POINT`
-                // reparse points, not `IO_REPARSE_TAG_SYMLINK`, and
-                // `FileType::is_symlink()` returns false for them.
-                // Catch every reparse point via the file-attribute
-                // bit so a junction can't sneak the patch out of the
-                // package directory.
-                #[cfg(windows)]
-                {
-                    use std::os::windows::fs::MetadataExt;
-                    const FILE_ATTRIBUTE_REPARSE_POINT: u32 = 0x0400;
-                    if meta.file_attributes() & FILE_ATTRIBUTE_REPARSE_POINT != 0 {
-                        return Err(format!("{}", cursor.display()));
-                    }
-                }
-            }
-            Err(ref e) if e.kind() == std::io::ErrorKind::NotFound => break,
-            Err(e) => return Err(format!("stat {}: {e}", cursor.display())),
-        }
-    }
-    Ok(())
-}
-
-fn apply_multi_file_patch(pkg_dir: &Path, patch_text: &str) -> Result<(), String> {
-    let sections = split_patch_sections(patch_text);
-    if sections.is_empty() {
-        return Err("patch contained no `diff --git` sections".to_string());
-    }
-    for section in sections {
-        let rel = section
-            .rel_path
-            .as_ref()
-            .ok_or_else(|| "patch section missing file path".to_string())?;
-        // Refuse patch headers that escape the package directory.
-        // A hostile diff with `b/../../etc/shadow` as the target
-        // would otherwise let the patch step overwrite or delete
-        // files outside the installed package. Same rules we apply
-        // to tar entries over in aube-store (no absolute, no drive
-        // prefix, no `..`, no backslash, no NUL).
-        if !is_safe_rel_component(rel) {
-            return Err(format!("patch file path escapes package: {rel:?}"));
-        }
-        // Walk every parent component of the target on disk and refuse
-        // to follow any symlink or junction. Without this guard, a
-        // package that planted a directory link inside its own tree
-        // (or a workspace where the user has a symlinked dep dir)
-        // would let `pkg_dir.join(rel)` resolve through the link, and
-        // `atomic_write` would overwrite a file outside `pkg_dir`.
-        // CVE-2018-1000156 (GNU patch) class.
-        if let Err(e) = ensure_no_symlink_in_chain(pkg_dir, rel) {
-            return Err(format!("patch target contains symlink: {e}"));
-        }
-        let target = pkg_dir.join(rel);
-        let original = if target.exists() {
-            std::fs::read_to_string(&target)
-                .map_err(|e| format!("failed to read {}: {e}", target.display()))?
-        } else {
-            String::new()
-        };
-        // `+++ /dev/null` means the patch deletes the file. Skip diffy
-        // entirely — `diffy::apply` would otherwise produce an empty
-        // string and we'd write a zero-byte file in place of the
-        // original, leaving `require('./removed')` resolving to an
-        // empty module instead of the expected `MODULE_NOT_FOUND`.
-        if section.is_deletion {
-            if target.exists() {
-                std::fs::remove_file(&target)
-                    .map_err(|e| format!("failed to remove {}: {e}", target.display()))?;
-            }
-            continue;
-        }
-        // git-style patches always use LF line endings, but published
-        // tarballs frequently ship files with CRLF (Windows editors,
-        // `core.autocrlf=true` checkouts). Diffy is byte-exact and
-        // refuses to match CRLF context against LF hunk lines, so we
-        // normalize the original to LF before applying and restore the
-        // CRLF on write. pnpm's patch applier does the same thing.
-        let was_crlf = original.contains("\r\n");
-        let normalized = if was_crlf {
-            original.replace("\r\n", "\n")
-        } else {
-            original
-        };
-        let parsed = diffy::Patch::from_str(&section.body)
-            .map_err(|e| format!("failed to parse patch for {rel}: {e}"))?;
-        let patched_lf = diffy::apply(&normalized, &parsed)
-            .map_err(|e| format!("failed to apply patch for {rel}: {e}"))?;
-        let patched = if was_crlf {
-            // Promote bare `\n` to `\r\n`, then collapse any `\r\r\n`
-            // back so a patch line containing a literal `\r` byte (rare
-            // but legal for binary-ish text) doesn't gain a second CR.
-            patched_lf.replace('\n', "\r\n").replace("\r\r\n", "\r\n")
-        } else {
-            patched_lf
-        };
-        // Break any reflink/hardlink to the global store before
-        // writing the patched bytes — otherwise we'd silently mutate
-        // every other project sharing this CAS file. Stage the write
-        // through a sibling tempfile and `rename` into place so a
-        // crash or Ctrl-C mid-patch cannot leave the package with
-        // the original file unlinked and no replacement written.
-        // POSIX `rename(2)` atomically replaces the destination, so
-        // no pre-removal is needed and removing first would create
-        // the exact TOCTOU window the rename is supposed to close.
-        // Windows `MoveFileExW` fails when the destination exists,
-        // so the unlink is gated behind `cfg(windows)`.
-        #[cfg(windows)]
-        {
-            if target.exists() {
-                std::fs::remove_file(&target)
-                    .map_err(|e| format!("failed to unlink {}: {e}", target.display()))?;
-            }
-        }
-        aube_util::fs_atomic::atomic_write(&target, patched.as_bytes()).map_err(|e| {
-            format!(
-                "failed to write patched file into place {}: {e}",
-                target.display()
-            )
-        })?;
-    }
-    Ok(())
-}
-
-struct PatchSection {
-    rel_path: Option<String>,
-    /// Single-file unified diff body — `diffy::Patch::from_str` parses
-    /// this directly. Always begins with `--- ` so the diffy parser
-    /// finds its anchor.
-    body: String,
-    /// `+++ /dev/null` was seen in the header — the patch deletes this
-    /// file, so the linker should `remove_file` instead of writing
-    /// patched bytes (which `diffy::apply` would emit as an empty
-    /// string).
-    is_deletion: bool,
-}
-
-/// Split a git-style multi-file patch into one section per file.
-/// We look for `diff --git a/<path> b/<path>` markers, pull the path
-/// out of the `b/...` half (post-edit name), and capture everything
-/// from the next `--- ` line until the following `diff --git ` (or
-/// EOF) as the diffy-compatible body.
-fn parse_diff_git_b_path(rest: &str) -> Option<String> {
-    if let Some(after) = rest.strip_prefix("\"a/") {
-        let end_a = after.find("\" \"b/")?;
-        let after_b = &after[end_a + 5..];
-        let close = after_b.rfind('"')?;
-        return unescape_git_quoted(&after_b[..close]);
-    }
-    let body = rest.strip_prefix("a/")?;
-    let mut search_from = 0;
-    while let Some(rel) = body[search_from..].find(" b/") {
-        let abs = search_from + rel;
-        let path_a = &body[..abs];
-        let path_b = &body[abs + 3..];
-        if path_a == path_b {
-            return Some(path_b.to_string());
-        }
-        search_from = abs + 1;
-    }
-    body.find(" b/").map(|i| body[i + 3..].to_string())
-}
-
-fn unescape_git_quoted(s: &str) -> Option<String> {
-    let bytes = s.as_bytes();
-    let mut out: Vec<u8> = Vec::with_capacity(bytes.len());
-    let mut i = 0;
-    while i < bytes.len() {
-        if bytes[i] != b'\\' {
-            out.push(bytes[i]);
-            i += 1;
-            continue;
-        }
-        if i + 1 >= bytes.len() {
-            return None;
-        }
-        match bytes[i + 1] {
-            b'\\' => {
-                out.push(b'\\');
-                i += 2;
-            }
-            b'"' => {
-                out.push(b'"');
-                i += 2;
-            }
-            b'n' => {
-                out.push(b'\n');
-                i += 2;
-            }
-            b't' => {
-                out.push(b'\t');
-                i += 2;
-            }
-            b'r' => {
-                out.push(b'\r');
-                i += 2;
-            }
-            b'a' => {
-                out.push(0x07);
-                i += 2;
-            }
-            b'b' => {
-                out.push(0x08);
-                i += 2;
-            }
-            b'f' => {
-                out.push(0x0C);
-                i += 2;
-            }
-            b'v' => {
-                out.push(0x0B);
-                i += 2;
-            }
-            d0 @ b'0'..=b'3'
-                if i + 3 < bytes.len()
-                    && (b'0'..=b'7').contains(&bytes[i + 2])
-                    && (b'0'..=b'7').contains(&bytes[i + 3]) =>
-            {
-                let n = ((d0 - b'0') << 6) | ((bytes[i + 2] - b'0') << 3) | (bytes[i + 3] - b'0');
-                out.push(n);
-                i += 4;
-            }
-            _ => return None,
-        }
-    }
-    String::from_utf8(out).ok()
-}
-
-fn split_patch_sections(text: &str) -> Vec<PatchSection> {
-    let mut out: Vec<PatchSection> = Vec::new();
-    let mut current_path: Option<String> = None;
-    let mut body = String::new();
-    let mut in_body = false;
-    let mut is_deletion = false;
-
-    let flush = |out: &mut Vec<PatchSection>,
-                 path: &mut Option<String>,
-                 body: &mut String,
-                 is_deletion: &mut bool| {
-        if !body.is_empty() || *is_deletion {
-            out.push(PatchSection {
-                rel_path: path.take(),
-                body: std::mem::take(body),
-                is_deletion: std::mem::replace(is_deletion, false),
-            });
-        } else {
-            *path = None;
-        }
-    };
-
-    for line in text.split_inclusive('\n') {
-        let stripped = line.trim_end_matches(['\n', '\r']);
-        if let Some(rest) = stripped.strip_prefix("diff --git ") {
-            // New file boundary — flush whatever we were collecting.
-            flush(&mut out, &mut current_path, &mut body, &mut is_deletion);
-            in_body = false;
-            // Parse `a/<path> b/<path>` and prefer the post-edit
-            // (`b/`) path so renames land on the new name.
-            current_path = parse_diff_git_b_path(rest);
-            continue;
-        }
-        if !in_body {
-            if stripped.starts_with("--- ") {
-                in_body = true;
-                // Rewrite `--- /dev/null` (file addition) to `--- a/<path>`
-                // so diffy's parser still gets a valid header. The
-                // original file content we feed `diffy::apply` is empty
-                // for additions, which is what diffy expects.
-                if stripped == "--- /dev/null"
-                    && let Some(rel) = current_path.as_deref()
-                {
-                    body.push_str(&format!("--- a/{rel}\n"));
-                } else {
-                    body.push_str(stripped);
-                    body.push('\n');
-                }
-            }
-            // Skip git's `index ...` / `new file mode ...` /
-            // `similarity index ...` decorations — diffy doesn't
-            // understand them and they aren't needed once we know
-            // the target path.
-            continue;
-        }
-        if stripped == "+++ /dev/null" {
-            // File deletion — note it and drop this header line. The
-            // linker will `remove_file` and skip the diffy apply path
-            // entirely, so the rest of the body (the hunk that empties
-            // the file) is intentionally discarded.
-            is_deletion = true;
-            continue;
-        }
-        body.push_str(stripped);
-        body.push('\n');
-    }
-    flush(&mut out, &mut current_path, &mut body, &mut is_deletion);
-    out
-}
-
-#[cfg(test)]
-mod importer_classification_tests {
-    use super::is_physical_importer;
-
-    #[test]
-    fn root_is_physical() {
-        assert!(is_physical_importer("."));
-    }
-
-    #[test]
-    fn workspace_paths_are_physical() {
-        assert!(is_physical_importer("packages/dev/core"));
-        assert!(is_physical_importer("apps/web"));
-        assert!(is_physical_importer("libs/@scope/name"));
-    }
-
-    #[test]
-    fn nested_peer_context_paths_are_virtual() {
-        // pnpm v9 emits these for every peer-resolution view reachable
-        // through the workspace symlink chain. They describe the graph,
-        // they are not directories to populate.
-        assert!(!is_physical_importer(
-            "packages/dev/addons/node_modules/@dev/core"
-        ));
-        assert!(!is_physical_importer(
-            "packages/a/node_modules/@s/b/node_modules/@s/c"
-        ));
-    }
-}
-
-#[cfg(test)]
-mod public_hoist_tests {
-    use super::*;
-
-    fn linker_with(patterns: &[&str]) -> Linker {
-        // Construct a Linker without touching disk: we only call
-        // `public_hoist_matches`, which never looks at `store` or
-        // `virtual_store`. A dummy store is acceptable because
-        // Store::clone is cheap and this test never invokes a method
-        // that would actually touch the CAS.
-        let store = Store::at(std::env::temp_dir().join("aube-public-hoist-test"));
-        let strs: Vec<String> = patterns.iter().map(|s| s.to_string()).collect();
-        Linker::new(&store, LinkStrategy::Copy).with_public_hoist_pattern(&strs)
-    }
-
-    #[test]
-    fn empty_pattern_matches_nothing() {
-        let l = linker_with(&[]);
-        assert!(!l.public_hoist_matches("react"));
-        assert!(!l.public_hoist_matches("eslint"));
-    }
-
-    #[test]
-    fn wildcard_matches_substring() {
-        let l = linker_with(&["*eslint*", "*prettier*"]);
-        assert!(l.public_hoist_matches("eslint"));
-        assert!(l.public_hoist_matches("eslint-plugin-react"));
-        assert!(l.public_hoist_matches("@typescript-eslint/parser"));
-        assert!(l.public_hoist_matches("prettier"));
-        assert!(!l.public_hoist_matches("react"));
-    }
-
-    #[test]
-    fn exact_name_match() {
-        let l = linker_with(&["react"]);
-        assert!(l.public_hoist_matches("react"));
-        assert!(!l.public_hoist_matches("react-dom"));
-    }
-
-    #[test]
-    fn negation_excludes_positive_match() {
-        let l = linker_with(&["*eslint*", "!eslint-config-*"]);
-        assert!(l.public_hoist_matches("eslint"));
-        assert!(l.public_hoist_matches("eslint-plugin-react"));
-        assert!(!l.public_hoist_matches("eslint-config-next"));
-    }
-
-    #[test]
-    fn case_insensitive() {
-        let l = linker_with(&["*ESLINT*"]);
-        assert!(l.public_hoist_matches("eslint"));
-        assert!(l.public_hoist_matches("ESLint"));
-    }
-
-    #[test]
-    fn invalid_patterns_are_silently_dropped() {
-        // `[` opens an unclosed character class — glob::Pattern::new
-        // rejects it; the builder skips the pattern instead of
-        // failing install. The accompanying valid pattern still
-        // matches.
-        let l = linker_with(&["[unterminated", "react"]);
-        assert!(l.public_hoist_matches("react"));
-        assert!(!l.public_hoist_matches("eslint"));
-    }
-}
-
-#[cfg(test)]
-mod patch_tests {
-    use super::*;
-
-    #[cfg(windows)]
-    #[test]
-    fn apply_multi_file_patch_refuses_to_follow_junction_outside_pkg() {
-        let outside = tempfile::tempdir().unwrap();
-        let pkg_root = tempfile::tempdir().unwrap();
-        let pkg = pkg_root.path().join("pkg");
-        std::fs::create_dir_all(&pkg).unwrap();
-        let escape = pkg.join("escape");
-        junction::create(outside.path(), &escape).unwrap();
-        let target = outside.path().join("victim.txt");
-        std::fs::write(&target, "untouched\n").unwrap();
-        let patch = "diff --git a/escape/victim.txt b/escape/victim.txt\n\
-                     --- a/escape/victim.txt\n\
-                     +++ b/escape/victim.txt\n\
-                     @@ -1 +1 @@\n\
-                     -untouched\n\
-                     +PWNED\n";
-        let result = apply_multi_file_patch(&pkg, patch);
-        assert!(result.is_err(), "patch must refuse junction-bearing rel");
-        let after = std::fs::read_to_string(&target).unwrap();
-        assert_eq!(after, "untouched\n");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn apply_multi_file_patch_refuses_to_follow_symlink_outside_pkg() {
-        let outside = tempfile::tempdir().unwrap();
-        let pkg_root = tempfile::tempdir().unwrap();
-        let pkg = pkg_root.path().join("pkg");
-        std::fs::create_dir_all(&pkg).unwrap();
-        let escape = pkg.join("escape");
-        std::os::unix::fs::symlink(outside.path(), &escape).unwrap();
-        let target = outside.path().join("victim.txt");
-        std::fs::write(&target, "untouched\n").unwrap();
-        let patch = "diff --git a/escape/victim.txt b/escape/victim.txt\n\
-                     --- a/escape/victim.txt\n\
-                     +++ b/escape/victim.txt\n\
-                     @@ -1 +1 @@\n\
-                     -untouched\n\
-                     +PWNED\n";
-        let result = apply_multi_file_patch(&pkg, patch);
-        assert!(result.is_err(), "patch must refuse symlink-bearing rel");
-        let after = std::fs::read_to_string(&target).unwrap();
-        assert_eq!(after, "untouched\n");
-    }
-
-    #[test]
-    fn round_trips_simple_patch() {
-        let dir = tempfile::tempdir().unwrap();
-        let pkg = dir.path().join("pkg");
-        std::fs::create_dir_all(&pkg).unwrap();
-        std::fs::write(pkg.join("index.js"), "module.exports = 'old';\n").unwrap();
-
-        let patch = "diff --git a/index.js b/index.js\n\
-                     --- a/index.js\n\
-                     +++ b/index.js\n\
-                     @@ -1 +1 @@\n\
-                     -module.exports = 'old';\n\
-                     +module.exports = 'new';\n";
-        apply_multi_file_patch(&pkg, patch).unwrap();
-        assert_eq!(
-            std::fs::read_to_string(pkg.join("index.js")).unwrap(),
-            "module.exports = 'new';\n"
-        );
-    }
-
-    #[test]
-    fn crlf_patch_path_does_not_carry_carriage_return() {
-        let patch = "diff --git a/index.js b/index.js\r\n\
-                     --- a/index.js\r\n\
-                     +++ b/index.js\r\n\
-                     @@ -1 +1 @@\r\n\
-                     -module.exports = 'old';\r\n\
-                     +module.exports = 'new';\r\n";
-        let sections = split_patch_sections(patch);
-        assert_eq!(sections.len(), 1);
-        assert_eq!(sections[0].rel_path.as_deref(), Some("index.js"));
-    }
-
-    #[test]
-    fn crlf_deletion_patch_recognized() {
-        let patch = "diff --git a/removed.js b/removed.js\r\n\
-                     deleted file mode 100644\r\n\
-                     --- a/removed.js\r\n\
-                     +++ /dev/null\r\n\
-                     @@ -1 +0,0 @@\r\n\
-                     -gone\r\n";
-        let sections = split_patch_sections(patch);
-        assert_eq!(sections.len(), 1);
-        assert!(sections[0].is_deletion);
-    }
-
-    #[test]
-    fn diff_git_path_with_space_b_substring() {
-        let patch = "diff --git a/a b/c.js b/a b/c.js\n\
-                     --- a/a b/c.js\n\
-                     +++ b/a b/c.js\n\
-                     @@ -1 +1 @@\n\
-                     -x\n\
-                     +y\n";
-        let sections = split_patch_sections(patch);
-        assert_eq!(sections.len(), 1);
-        assert_eq!(sections[0].rel_path.as_deref(), Some("a b/c.js"));
-    }
-
-    #[test]
-    fn diff_git_quoted_path_form() {
-        let patch = "diff --git \"a/path with spaces.js\" \"b/path with spaces.js\"\n\
-                     --- a/path with spaces.js\n\
-                     +++ b/path with spaces.js\n\
-                     @@ -1 +1 @@\n\
-                     -x\n\
-                     +y\n";
-        let sections = split_patch_sections(patch);
-        assert_eq!(sections.len(), 1);
-        assert_eq!(sections[0].rel_path.as_deref(), Some("path with spaces.js"));
-    }
-
-    #[test]
-    fn applies_lf_patch_against_crlf_file() {
-        // Tarballs published from Windows editors ship CRLF text. pnpm
-        // / git emit LF-only patches even against those files. Diffy is
-        // byte-exact, so the apply path normalizes CRLF -> LF before
-        // matching and restores CRLF on write.
-        let dir = tempfile::tempdir().unwrap();
-        let pkg = dir.path().join("pkg");
-        std::fs::create_dir_all(&pkg).unwrap();
-        std::fs::write(pkg.join("a.txt"), b"one\r\ntwo\r\nthree\r\n").unwrap();
-
-        let patch = "diff --git a/a.txt b/a.txt\n\
-                     --- a/a.txt\n\
-                     +++ b/a.txt\n\
-                     @@ -1,3 +1,3 @@\n\
-                     \x20one\n\
-                     -two\n\
-                     +TWO\n\
-                     \x20three\n";
-        apply_multi_file_patch(&pkg, patch).unwrap();
-        let bytes = std::fs::read(pkg.join("a.txt")).unwrap();
-        assert_eq!(bytes, b"one\r\nTWO\r\nthree\r\n");
-    }
-
-    #[test]
-    fn crlf_restore_preserves_embedded_cr_byte() {
-        // A patch line that adds a literal `\r` byte mid-line must not
-        // gain a second `\r` when we re-CRLF the output. Naive
-        // `replace('\n', "\r\n")` would turn `\r\n` into `\r\r\n`; the
-        // `\r\r\n` collapse undoes that.
-        let dir = tempfile::tempdir().unwrap();
-        let pkg = dir.path().join("pkg");
-        std::fs::create_dir_all(&pkg).unwrap();
-        std::fs::write(pkg.join("a.txt"), b"one\r\ntwo\r\n").unwrap();
-        let patch = "diff --git a/a.txt b/a.txt\n\
-                     --- a/a.txt\n\
-                     +++ b/a.txt\n\
-                     @@ -1,2 +1,2 @@\n\
-                     -one\n\
-                     +has\rcr\n\
-                     \x20two\n";
-        apply_multi_file_patch(&pkg, patch).unwrap();
-        let bytes = std::fs::read(pkg.join("a.txt")).unwrap();
-        assert_eq!(bytes, b"has\rcr\r\ntwo\r\n");
-    }
-
-    #[test]
-    fn diff_git_quoted_path_unescapes_git_escapes() {
-        let path = parse_diff_git_b_path(r#""a/foo\".js" "b/foo\".js""#).expect("quoted parse");
-        assert_eq!(path, "foo\".js");
-        let path = parse_diff_git_b_path(r#""a/back\\slash.js" "b/back\\slash.js""#)
-            .expect("backslash parse");
-        assert_eq!(path, "back\\slash.js");
-        let path = parse_diff_git_b_path("\"a/caf\\303\\251.js\" \"b/caf\\303\\251.js\"")
-            .expect("octal parse");
-        assert_eq!(path, "café.js");
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use aube_lockfile::{DepType, DirectDep, LockedPackage, LockfileGraph};
-    use aube_store::Store;
-
-    fn setup_store_with_files(dir: &Path) -> (Store, BTreeMap<String, aube_store::PackageIndex>) {
-        let store = Store::at(dir.join("store/files"));
-
-        let mut indices = BTreeMap::new();
-
-        // foo@1.0.0 with index.js
-        let foo_stored = store
-            .import_bytes(b"module.exports = 'foo';", false)
-            .unwrap();
-        let mut foo_index = PackageIndex::default();
-        foo_index.insert("index.js".to_string(), foo_stored);
-
-        // foo also has package.json
-        let foo_pkg = store
-            .import_bytes(b"{\"name\":\"foo\",\"version\":\"1.0.0\"}", false)
-            .unwrap();
-        foo_index.insert("package.json".to_string(), foo_pkg);
-        indices.insert("foo@1.0.0".to_string(), foo_index);
-
-        // bar@2.0.0 with index.js
-        let bar_stored = store
-            .import_bytes(b"module.exports = 'bar';", false)
-            .unwrap();
-        let mut bar_index = PackageIndex::default();
-        bar_index.insert("index.js".to_string(), bar_stored);
-        indices.insert("bar@2.0.0".to_string(), bar_index);
-
-        (store, indices)
-    }
-
-    fn make_graph() -> LockfileGraph {
-        let mut packages = BTreeMap::new();
-
-        let mut foo_deps = BTreeMap::new();
-        foo_deps.insert("bar".to_string(), "2.0.0".to_string());
-
-        packages.insert(
-            "foo@1.0.0".to_string(),
-            LockedPackage {
-                name: "foo".to_string(),
-                version: "1.0.0".to_string(),
-                integrity: None,
-                dependencies: foo_deps,
-                dep_path: "foo@1.0.0".to_string(),
-                ..Default::default()
-            },
-        );
-        packages.insert(
-            "bar@2.0.0".to_string(),
-            LockedPackage {
-                name: "bar".to_string(),
-                version: "2.0.0".to_string(),
-                integrity: None,
-                dependencies: BTreeMap::new(),
-                dep_path: "bar@2.0.0".to_string(),
-                ..Default::default()
-            },
-        );
-
-        let mut importers = BTreeMap::new();
-        importers.insert(
-            ".".to_string(),
-            vec![DirectDep {
-                name: "foo".to_string(),
-                dep_path: "foo@1.0.0".to_string(),
-                dep_type: DepType::Production,
-                specifier: None,
-            }],
-        );
-
-        LockfileGraph {
-            importers,
-            packages,
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn test_detect_strategy() {
-        let dir = tempfile::tempdir().unwrap();
-        let strategy = Linker::detect_strategy(dir.path());
-        // Probe returns `Hardlink` or `Copy`; `Reflink` is only
-        // reachable through explicit `packageImportMethod =
-        // clone`/`clone-or-copy`, so the match guards that contract.
-        match strategy {
-            LinkStrategy::Hardlink | LinkStrategy::Copy => {}
-            LinkStrategy::Reflink => panic!("detect_strategy must not return Reflink"),
-        }
-    }
-
-    #[test]
-    fn test_link_all_handles_self_referential_dep_at_different_version() {
-        // `react_ujs@3.3.0` (and other publish-script artifacts)
-        // declares its own name as a dep at a *different* version
-        // (`react_ujs: ^2.7.1`). The transitive-symlink pass would
-        // try to create a symlink at `node_modules/react_ujs`,
-        // which is exactly where the package's own files live —
-        // EEXIST. Skip self-name deps regardless of version so
-        // these install cleanly. `require('<self>')` from inside
-        // the package then resolves to its own files, matching how
-        // npm / pnpm / yarn end up after their hoisting passes.
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let store = Store::at(dir.path().join("store/files"));
-
-        let mut indices = BTreeMap::new();
-        let host_index_js = store.import_bytes(b"/* react_ujs 3.3.0 */", false).unwrap();
-        let host_pkg_json = store
-            .import_bytes(b"{\"name\":\"react_ujs\",\"version\":\"3.3.0\"}", false)
-            .unwrap();
-        let mut host_index = PackageIndex::default();
-        host_index.insert("index.js".to_string(), host_index_js);
-        host_index.insert("package.json".to_string(), host_pkg_json);
-        indices.insert("react_ujs@3.3.0".to_string(), host_index);
-
-        let mut host_deps = BTreeMap::new();
-        // Self-reference at a different version, the shape that
-        // triggered the EEXIST bug.
-        host_deps.insert("react_ujs".to_string(), "^2.7.1".to_string());
-
-        let mut packages = BTreeMap::new();
-        packages.insert(
-            "react_ujs@3.3.0".to_string(),
-            LockedPackage {
-                name: "react_ujs".to_string(),
-                version: "3.3.0".to_string(),
-                integrity: None,
-                dependencies: host_deps,
-                dep_path: "react_ujs@3.3.0".to_string(),
-                ..Default::default()
-            },
-        );
-
-        let mut importers = BTreeMap::new();
-        importers.insert(
-            ".".to_string(),
-            vec![DirectDep {
-                name: "react_ujs".to_string(),
-                dep_path: "react_ujs@3.3.0".to_string(),
-                dep_type: DepType::Production,
-                specifier: None,
-            }],
-        );
-
-        let graph = LockfileGraph {
-            importers,
-            packages,
-            ..Default::default()
-        };
-
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Copy, true);
-        let stats = linker
-            .link_all(&project_dir, &graph, &indices)
-            .expect("install must succeed despite self-named dep");
-        assert_eq!(stats.packages_linked, 1);
-        let host_index =
-            project_dir.join("node_modules/.aube/react_ujs@3.3.0/node_modules/react_ujs/index.js");
-        assert!(host_index.exists(), "host package files must be present");
-    }
-
-    #[test]
-    fn test_link_all_creates_pnpm_virtual_store() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Copy, true);
-        let graph = make_graph();
-
-        let stats = linker.link_all(&project_dir, &graph, &indices).unwrap();
-
-        // .aube virtual store should exist
-        assert!(project_dir.join("node_modules/.aube").exists());
-
-        // .aube/foo@1.0.0 should be a symlink to the global virtual store
-        let aube_foo = project_dir.join("node_modules/.aube/foo@1.0.0");
-        assert!(aube_foo.symlink_metadata().unwrap().is_symlink());
-
-        // foo@1.0.0 content should be accessible through the symlink
-        let foo_in_pnpm =
-            project_dir.join("node_modules/.aube/foo@1.0.0/node_modules/foo/index.js");
-        assert!(foo_in_pnpm.exists());
-        assert_eq!(
-            std::fs::read_to_string(&foo_in_pnpm).unwrap(),
-            "module.exports = 'foo';"
-        );
-
-        // bar@2.0.0 should also be accessible
-        let bar_in_pnpm =
-            project_dir.join("node_modules/.aube/bar@2.0.0/node_modules/bar/index.js");
-        assert!(bar_in_pnpm.exists());
-
-        assert_eq!(stats.packages_linked, 2);
-        assert!(stats.files_linked >= 3); // foo has 2 files, bar has 1
-    }
-
-    #[test]
-    fn test_link_file_fresh_reports_missing_cas_shard_and_invalidates_cache() {
-        // Reproduces endevco/aube#393: a partially corrupt CAS leaves the
-        // cached package index pointing at a missing shard. Materialize
-        // must distinguish "source CAS file missing" from a generic ENOENT
-        // and drop the stale index JSON so the next install re-imports
-        // the tarball.
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        // Persist foo's index so invalidate_cached_index has something
-        // to remove. Real installs save indices via the fetch path.
-        let foo_index = indices.get("foo@1.0.0").unwrap();
-        store.save_index("foo", "1.0.0", None, foo_index).unwrap();
-        let cached_path = store.index_dir().join("foo@1.0.0.json");
-        assert!(
-            cached_path.exists(),
-            "test setup: index cache must be written"
-        );
-
-        // Delete the CAS shard for foo's package.json (matches the
-        // failure mode in #393 where one shard is missing while others
-        // remain).
-        let pkgjson_store_path = foo_index.get("package.json").unwrap().store_path.clone();
-        std::fs::remove_file(&pkgjson_store_path).unwrap();
-
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Copy, true);
-        let graph = make_graph();
-        let err = linker
-            .link_all(&project_dir, &graph, &indices)
-            .expect_err("link must fail when a referenced CAS shard is gone");
-        assert!(
-            matches!(&err, Error::MissingStoreFile { rel_path, .. } if rel_path == "package.json"),
-            "expected MissingStoreFile {{ rel_path: \"package.json\" }}, got {err:?}"
-        );
-
-        // Side effect: cached index dropped, so the next install will
-        // miss load_index and re-fetch instead of looping on the same
-        // dead shard reference.
-        assert!(
-            !cached_path.exists(),
-            "stale index cache must be invalidated on MissingStoreFile"
-        );
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn test_link_file_fresh_hardlink_short_circuits_when_source_missing() {
-        // Hardlink path used to silently fall through to `std::fs::copy`
-        // on ENOENT and emit a misleading "hardlink failed, falling back
-        // to copy" trace, even though the real cause was the source
-        // shard going missing. Short-circuit returns MissingStoreFile
-        // directly so traces stay accurate.
-        let dir = tempfile::tempdir().unwrap();
-        let store = Store::at(dir.path().join("store/files"));
-        let stored = store.import_bytes(b"hello", false).unwrap();
-        // Capture the path before we move `stored` into link_file_fresh.
-        let store_path = stored.store_path.clone();
-        std::fs::remove_file(&store_path).unwrap();
-
-        let dst_dir = dir.path().join("dst");
-        std::fs::create_dir_all(&dst_dir).unwrap();
-        let dst = dst_dir.join("hello.txt");
-
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Hardlink, true);
-        let err = linker
-            .link_file_fresh(&stored, "hello.txt", &dst)
-            .expect_err("source missing must fail");
-        assert!(
-            matches!(
-                &err,
-                Error::MissingStoreFile { store_path: p, rel_path } if p == &store_path && rel_path == "hello.txt"
-            ),
-            "expected MissingStoreFile from Hardlink branch, got {err:?}"
-        );
-    }
-
-    #[test]
-    fn test_link_all_creates_top_level_entries() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let linker = Linker::new(&store, LinkStrategy::Copy);
-        let graph = make_graph();
-
-        let stats = linker.link_all(&project_dir, &graph, &indices).unwrap();
-
-        // Top-level foo/ should exist (it's a direct dep)
-        let foo_top = project_dir.join("node_modules/foo/index.js");
-        assert!(foo_top.exists());
-        assert_eq!(
-            std::fs::read_to_string(&foo_top).unwrap(),
-            "module.exports = 'foo';"
-        );
-
-        // bar should NOT be top-level (it's only a transitive dep)
-        let bar_top = project_dir.join("node_modules/bar/index.js");
-        assert!(!bar_top.exists());
-
-        assert_eq!(stats.top_level_linked, 1);
-    }
-
-    #[test]
-    fn test_link_all_transitive_symlinks() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let linker = Linker::new(&store, LinkStrategy::Copy);
-        let graph = make_graph();
-
-        linker.link_all(&project_dir, &graph, &indices).unwrap();
-
-        // foo's node_modules/bar should be a symlink (inside the global virtual store)
-        // The path resolves through the .aube symlink into the global store
-        let bar_symlink = project_dir.join("node_modules/.aube/foo@1.0.0/node_modules/bar");
-        assert!(bar_symlink.symlink_metadata().unwrap().is_symlink());
-    }
-
-    #[test]
-    fn test_link_all_cleans_existing_node_modules() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        let nm = project_dir.join("node_modules");
-        std::fs::create_dir_all(&nm).unwrap();
-        std::fs::write(nm.join("stale-file.txt"), "old").unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let linker = Linker::new(&store, LinkStrategy::Copy);
-        let graph = make_graph();
-
-        linker.link_all(&project_dir, &graph, &indices).unwrap();
-
-        // Old file should be gone
-        assert!(!nm.join("stale-file.txt").exists());
-        // New structure should exist
-        assert!(nm.join(".aube").exists());
-    }
-
-    #[test]
-    fn test_link_all_nested_node_modules_for_direct_deps() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let linker = Linker::new(&store, LinkStrategy::Copy);
-        let graph = make_graph();
-
-        linker.link_all(&project_dir, &graph, &indices).unwrap();
-
-        // foo is a direct dep with bar as a transitive dep.
-        // The top-level node_modules/foo is a symlink to .aube/foo@1.0.0/node_modules/foo,
-        // and bar lives as a sibling at .aube/foo@1.0.0/node_modules/bar (also a symlink
-        // pointing to .aube/bar@2.0.0/node_modules/bar). Node's directory walk from inside
-        // foo finds bar this way without aube creating any nested node_modules.
-        let foo_link = project_dir.join("node_modules/foo");
-        assert!(foo_link.symlink_metadata().unwrap().is_symlink());
-        let bar_sibling = project_dir.join("node_modules/.aube/foo@1.0.0/node_modules/bar");
-        assert!(bar_sibling.symlink_metadata().unwrap().is_symlink());
-    }
-
-    #[test]
-    fn test_global_virtual_store_is_populated() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let virtual_store = store.virtual_store_dir();
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Copy, true);
-        let graph = make_graph();
-
-        linker.link_all(&project_dir, &graph, &indices).unwrap();
-
-        // Global virtual store should contain materialized packages
-        let foo_global = virtual_store.join("foo@1.0.0/node_modules/foo/index.js");
-        assert!(foo_global.exists());
-        assert_eq!(
-            std::fs::read_to_string(&foo_global).unwrap(),
-            "module.exports = 'foo';"
-        );
-
-        let bar_global = virtual_store.join("bar@2.0.0/node_modules/bar/index.js");
-        assert!(bar_global.exists());
-    }
-
-    #[test]
-    fn test_global_virtual_store_gets_hidden_hoist() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let virtual_store = store.virtual_store_dir();
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Copy, true);
-        let mut graph = make_graph();
-        graph
-            .packages
-            .get_mut("foo@1.0.0")
-            .unwrap()
-            .dependencies
-            .clear();
-
-        linker.link_all(&project_dir, &graph, &indices).unwrap();
-
-        let project_hidden = project_dir.join("node_modules/.aube/node_modules/bar");
-        assert!(project_hidden.symlink_metadata().unwrap().is_symlink());
-
-        let global_hidden = virtual_store.join("node_modules/bar");
-        assert!(global_hidden.symlink_metadata().unwrap().is_symlink());
-
-        let from_real_store = virtual_store.join("foo@1.0.0/node_modules/bar/index.js");
-        assert!(
-            !from_real_store.exists(),
-            "bar is not a declared sibling of foo in this fixture"
-        );
-        let fallback = virtual_store.join("node_modules/bar/index.js");
-        assert_eq!(
-            std::fs::read_to_string(fallback).unwrap(),
-            "module.exports = 'bar';"
-        );
-    }
-
-    #[test]
-    fn test_global_virtual_store_hidden_hoist_prunes_only_dead_entries() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let virtual_store = store.virtual_store_dir();
-        let hidden = virtual_store.join("node_modules");
-        std::fs::create_dir_all(&hidden).unwrap();
-        let dotfile = hidden.join(".sentinel");
-        std::fs::write(&dotfile, "shared").unwrap();
-        let stale = hidden.join("stale");
-        std::fs::write(&stale, "old").unwrap();
-        let stale_scope = hidden.join("@stale-scope");
-        std::fs::write(&stale_scope, "old").unwrap();
-        let external_target = virtual_store.join("external@1.0.0/node_modules/external");
-        std::fs::create_dir_all(&external_target).unwrap();
-        let external_link = hidden.join("external");
-        sys::create_dir_link(
-            &pathdiff::diff_paths(&external_target, &hidden).unwrap(),
-            &external_link,
-        )
-        .unwrap();
-        let dead_link = hidden.join("dead");
-        sys::create_dir_link(
-            Path::new("../missing@1.0.0/node_modules/missing"),
-            &dead_link,
-        )
-        .unwrap();
-
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Copy, true);
-        linker
-            .link_all(&project_dir, &make_graph(), &indices)
-            .unwrap();
-
-        assert_eq!(std::fs::read_to_string(dotfile).unwrap(), "shared");
-        assert!(!stale.exists());
-        assert!(stale_scope.symlink_metadata().is_err());
-        assert!(external_link.symlink_metadata().unwrap().is_symlink());
-        assert!(dead_link.symlink_metadata().is_err());
-        assert!(hidden.join("bar").symlink_metadata().unwrap().is_symlink());
-    }
-
-    #[test]
-    fn test_global_virtual_store_hidden_hoist_disabled_keeps_live_shared_links() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let virtual_store = store.virtual_store_dir();
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Copy, true);
-        linker
-            .link_all(&project_dir, &make_graph(), &indices)
-            .unwrap();
-
-        let global_hidden = virtual_store.join("node_modules/bar");
-        assert!(global_hidden.symlink_metadata().unwrap().is_symlink());
-
-        Linker::new_with_gvs(&store, LinkStrategy::Copy, true)
-            .with_hoist(false)
-            .link_all(&project_dir, &make_graph(), &indices)
-            .unwrap();
-
-        assert!(global_hidden.symlink_metadata().unwrap().is_symlink());
-    }
-
-    #[test]
-    fn test_second_install_reuses_global_store() {
-        let dir = tempfile::tempdir().unwrap();
-
-        let (store, indices) = setup_store_with_files(dir.path());
-        let linker = Linker::new_with_gvs(&store, LinkStrategy::Copy, true);
-        let graph = make_graph();
-
-        // First install
-        let project1 = dir.path().join("project1");
-        std::fs::create_dir_all(&project1).unwrap();
-        let stats1 = linker.link_all(&project1, &graph, &indices).unwrap();
-        assert_eq!(stats1.packages_linked, 2);
-        assert_eq!(stats1.packages_cached, 0);
-
-        // Second install with same deps — should reuse global virtual store
-        let project2 = dir.path().join("project2");
-        std::fs::create_dir_all(&project2).unwrap();
-        let stats2 = linker.link_all(&project2, &graph, &indices).unwrap();
-        assert_eq!(stats2.packages_linked, 0);
-        assert_eq!(stats2.packages_cached, 2);
-        assert_eq!(stats2.files_linked, 0); // no CAS linking needed
-
-        // Both projects should work
-        let foo1 = project1.join("node_modules/foo/index.js");
-        let foo2 = project2.join("node_modules/foo/index.js");
-        assert!(foo1.exists());
-        assert!(foo2.exists());
-        assert_eq!(
-            std::fs::read_to_string(&foo1).unwrap(),
-            std::fs::read_to_string(&foo2).unwrap()
-        );
-    }
-
-    /// Regression: a version bump keeps the same top-level name
-    /// (`foo`) but must repoint `node_modules/foo` at the new
-    /// `.aube/foo@<new>` entry. The old `.aube/foo@<old>/` is left
-    /// on disk (no one sweeps the virtual store by name), so a
-    /// plain `path.exists()` check would see a still-resolving
-    /// stale symlink and keep it. The target-aware
-    /// `reconcile_top_level_link` compares the expected target
-    /// string and rewrites the link.
-    #[test]
-    fn test_link_all_repoints_symlink_after_version_bump() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-        let store = Store::at(dir.path().join("store/files"));
-
-        // Install 1: foo@1.0.0 as the root's direct dep.
-        let mut indices_v1 = BTreeMap::new();
-        let foo_v1 = store
-            .import_bytes(b"module.exports = 'foo@1';", false)
-            .unwrap();
-        let mut foo_v1_index = PackageIndex::default();
-        foo_v1_index.insert("index.js".to_string(), foo_v1);
-        indices_v1.insert("foo@1.0.0".to_string(), foo_v1_index);
-
-        let mut graph_v1 = LockfileGraph::default();
-        graph_v1.packages.insert(
-            "foo@1.0.0".to_string(),
-            LockedPackage {
-                name: "foo".to_string(),
-                version: "1.0.0".to_string(),
-                dep_path: "foo@1.0.0".to_string(),
-                ..Default::default()
-            },
-        );
-        graph_v1.importers.insert(
-            ".".to_string(),
-            vec![DirectDep {
-                name: "foo".to_string(),
-                dep_path: "foo@1.0.0".to_string(),
-                dep_type: DepType::Production,
-                specifier: None,
-            }],
-        );
-
-        let linker = Linker::new(&store, LinkStrategy::Copy);
-        linker
-            .link_all(&project_dir, &graph_v1, &indices_v1)
-            .unwrap();
-        let foo_link = project_dir.join("node_modules/foo");
-        assert!(foo_link.symlink_metadata().unwrap().is_symlink());
-        assert_eq!(
-            std::fs::read_to_string(foo_link.join("index.js")).unwrap(),
-            "module.exports = 'foo@1';"
-        );
-
-        // Install 2: foo upgraded to 2.0.0. The `.aube/foo@1.0.0/`
-        // tree stays on disk (nothing prunes the virtual store by
-        // name), so the old `node_modules/foo` symlink still
-        // resolves — a naive "does the target exist?" check would
-        // keep it.
-        let mut indices_v2 = BTreeMap::new();
-        let foo_v2 = store
-            .import_bytes(b"module.exports = 'foo@2';", false)
-            .unwrap();
-        let mut foo_v2_index = PackageIndex::default();
-        foo_v2_index.insert("index.js".to_string(), foo_v2);
-        indices_v2.insert("foo@2.0.0".to_string(), foo_v2_index);
-
-        let mut graph_v2 = LockfileGraph::default();
-        graph_v2.packages.insert(
-            "foo@2.0.0".to_string(),
-            LockedPackage {
-                name: "foo".to_string(),
-                version: "2.0.0".to_string(),
-                dep_path: "foo@2.0.0".to_string(),
-                ..Default::default()
-            },
-        );
-        graph_v2.importers.insert(
-            ".".to_string(),
-            vec![DirectDep {
-                name: "foo".to_string(),
-                dep_path: "foo@2.0.0".to_string(),
-                dep_type: DepType::Production,
-                specifier: None,
-            }],
-        );
-        linker
-            .link_all(&project_dir, &graph_v2, &indices_v2)
-            .unwrap();
-
-        // The top-level symlink must now resolve to foo@2.0.0's
-        // bytes, not foo@1.0.0's.
-        assert_eq!(
-            std::fs::read_to_string(project_dir.join("node_modules/foo/index.js")).unwrap(),
-            "module.exports = 'foo@2';"
-        );
-    }
-
-    /// Regression: `shamefully_hoist` hoists transitive deps to the
-    /// top-level `node_modules/<name>`. When the hoisted version
-    /// changes between installs (transitive bump), the previous
-    /// implementation kept the stale symlink because
-    /// `keep_or_reclaim_broken_symlink` only checked "does target
-    /// resolve?" and the old `.aube/<old-dep-path>/` was still on
-    /// disk. `reconcile_top_level_link` + the explicit
-    /// direct-dep/claimed tracking in `hoist_remaining_into` together
-    /// fix this.
-    #[test]
-    fn test_shamefully_hoist_repoints_after_transitive_version_bump() {
-        let dir = tempfile::tempdir().unwrap();
-        let project_dir = dir.path().join("project");
-        std::fs::create_dir_all(&project_dir).unwrap();
-        let store = Store::at(dir.path().join("store/files"));
-
-        // Install 1: root → bar@1.0.0 → foo@1.0.0 (transitive).
-        let foo_v1 = store
-            .import_bytes(b"module.exports = 'foo@1';", false)
-            .unwrap();
-        let mut foo_v1_idx = PackageIndex::default();
-        foo_v1_idx.insert("index.js".to_string(), foo_v1);
-        let bar_v1 = store
-            .import_bytes(b"module.exports = 'bar@1';", false)
-            .unwrap();
-        let mut bar_v1_idx = PackageIndex::default();
-        bar_v1_idx.insert("index.js".to_string(), bar_v1);
-        let mut indices_v1 = BTreeMap::new();
-        indices_v1.insert("foo@1.0.0".to_string(), foo_v1_idx);
-        indices_v1.insert("bar@1.0.0".to_string(), bar_v1_idx);
-
-        let mut graph_v1 = LockfileGraph::default();
-        let mut bar_deps_v1 = BTreeMap::new();
-        bar_deps_v1.insert("foo".to_string(), "1.0.0".to_string());
-        graph_v1.packages.insert(
-            "bar@1.0.0".to_string(),
-            LockedPackage {
-                name: "bar".to_string(),
-                version: "1.0.0".to_string(),
-                dep_path: "bar@1.0.0".to_string(),
-                dependencies: bar_deps_v1,
-                ..Default::default()
-            },
-        );
-        graph_v1.packages.insert(
-            "foo@1.0.0".to_string(),
-            LockedPackage {
-                name: "foo".to_string(),
-                version: "1.0.0".to_string(),
-                dep_path: "foo@1.0.0".to_string(),
-                ..Default::default()
-            },
-        );
-        graph_v1.importers.insert(
-            ".".to_string(),
-            vec![DirectDep {
-                name: "bar".to_string(),
-                dep_path: "bar@1.0.0".to_string(),
-                dep_type: DepType::Production,
-                specifier: None,
-            }],
-        );
-
-        let linker = Linker::new(&store, LinkStrategy::Copy).with_shamefully_hoist(true);
-        linker
-            .link_all(&project_dir, &graph_v1, &indices_v1)
-            .unwrap();
-        assert_eq!(
-            std::fs::read_to_string(project_dir.join("node_modules/foo/index.js")).unwrap(),
-            "module.exports = 'foo@1';",
-            "install 1 should hoist foo@1.0.0"
-        );
-
-        // Install 2: bar@1.0.0 → foo@2.0.0 (transitive bump). The
-        // stale `.aube/foo@1.0.0/` tree is still on disk (nothing
-        // sweeps the virtual store by name), so the old hoisted
-        // symlink would still resolve — the old `exists?` check
-        // would silently keep it.
-        let foo_v2 = store
-            .import_bytes(b"module.exports = 'foo@2';", false)
-            .unwrap();
-        let mut foo_v2_idx = PackageIndex::default();
-        foo_v2_idx.insert("index.js".to_string(), foo_v2);
-        let mut indices_v2 = BTreeMap::new();
-        // Reuse bar's materialized index from v1.
-        let bar_v1_for_v2 = store
-            .import_bytes(b"module.exports = 'bar@1';", false)
-            .unwrap();
-        let mut bar_v1_idx_v2 = PackageIndex::default();
-        bar_v1_idx_v2.insert("index.js".to_string(), bar_v1_for_v2);
-        indices_v2.insert("bar@1.0.0".to_string(), bar_v1_idx_v2);
-        indices_v2.insert("foo@2.0.0".to_string(), foo_v2_idx);
-
-        let mut graph_v2 = LockfileGraph::default();
-        let mut bar_deps_v2 = BTreeMap::new();
-        bar_deps_v2.insert("foo".to_string(), "2.0.0".to_string());
-        graph_v2.packages.insert(
-            "bar@1.0.0".to_string(),
-            LockedPackage {
-                name: "bar".to_string(),
-                version: "1.0.0".to_string(),
-                dep_path: "bar@1.0.0".to_string(),
-                dependencies: bar_deps_v2,
-                ..Default::default()
-            },
-        );
-        graph_v2.packages.insert(
-            "foo@2.0.0".to_string(),
-            LockedPackage {
-                name: "foo".to_string(),
-                version: "2.0.0".to_string(),
-                dep_path: "foo@2.0.0".to_string(),
-                ..Default::default()
-            },
-        );
-        graph_v2.importers.insert(
-            ".".to_string(),
-            vec![DirectDep {
-                name: "bar".to_string(),
-                dep_path: "bar@1.0.0".to_string(),
-                dep_type: DepType::Production,
-                specifier: None,
-            }],
-        );
-
-        linker
-            .link_all(&project_dir, &graph_v2, &indices_v2)
-            .unwrap();
-        assert_eq!(
-            std::fs::read_to_string(project_dir.join("node_modules/foo/index.js")).unwrap(),
-            "module.exports = 'foo@2';",
-            "install 2 should repoint the hoisted symlink to foo@2.0.0"
-        );
-    }
-
-    // ---------------------------------------------------------------
-    // `validate_index_key` rejects every shape of index key that
-    // would make `base.join(key)` escape `base`. Primary defence is
-    // in `aube-store::import_tarball`; this is the last-chance guard
-    // before the linker actually writes to disk.
-    // ---------------------------------------------------------------
-
-    #[test]
-    fn validate_index_key_accepts_normal_keys() {
-        validate_index_key("index.js").unwrap();
-        validate_index_key("lib/sub/a.js").unwrap();
-        validate_index_key("package.json").unwrap();
-        validate_index_key("a/b/c/d/e/f.js").unwrap();
-    }
-
-    #[cfg(not(windows))]
-    #[test]
-    fn validate_index_key_accepts_posix_colon_filename() {
-        validate_index_key("dist/__mocks__/package-json:version.d.ts").unwrap();
-    }
-
-    #[test]
-    fn validate_index_key_rejects_empty() {
-        assert!(matches!(
-            validate_index_key(""),
-            Err(Error::UnsafeIndexKey(_))
-        ));
-    }
-
-    #[test]
-    fn validate_index_key_rejects_leading_slash() {
-        assert!(matches!(
-            validate_index_key("/etc/passwd"),
-            Err(Error::UnsafeIndexKey(_))
-        ));
-        assert!(matches!(
-            validate_index_key("\\evil"),
-            Err(Error::UnsafeIndexKey(_))
-        ));
-    }
-
-    #[test]
-    fn validate_index_key_rejects_parent_dir() {
-        assert!(matches!(
-            validate_index_key("../../etc/passwd"),
-            Err(Error::UnsafeIndexKey(_))
-        ));
-        assert!(matches!(
-            validate_index_key("lib/../../../etc"),
-            Err(Error::UnsafeIndexKey(_))
-        ));
-    }
-
-    #[test]
-    fn validate_index_key_rejects_nul_and_backslash() {
-        assert!(matches!(
-            validate_index_key("lib\0evil"),
-            Err(Error::UnsafeIndexKey(_))
-        ));
-        assert!(matches!(
-            validate_index_key("lib\\..\\etc"),
-            Err(Error::UnsafeIndexKey(_))
-        ));
-    }
-
-    #[cfg(windows)]
-    #[test]
-    fn validate_index_key_rejects_windows_drive() {
-        assert!(matches!(
-            validate_index_key("C:Windows"),
-            Err(Error::UnsafeIndexKey(_))
-        ));
-    }
 }
